% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_nds_group.R
\name{morph_nds_group}
\alias{morph_nds_group}
\title{Morphometrics classification between GUs}
\usage{
morph_nds_group(
  nodes,
  focus = c("clust", "kmeans"),
  gu.types = "all",
  nb.centers = 1,
  out.dir = "_out",
  out.data = c("mbrshp", "plot")
)
}
\arguments{
\item{nodes}{Dataframe of nodes}

\item{focus}{Type of grouping, hierachical clustering ("clust") or Kmeans ("kmeans").
By default, c("clust", "kmeans")}

\item{gu.types}{Classes of nodes that will be clustered, a vector of characters or a character.
By default "all"}

\item{nb.centers}{Number of clusters, uniquely for Kmeans. By default 1 (unique cluster)}

\item{out.dir}{Name of output folder}

\item{out.data}{Type of data returned.
If "mbrshp" return a dataframe of nodes with their clustering and image path.
If "plot" return a "kmeans" or create a plot. By default c("mbrshp", "plot")}
}
\value{
Depending on the focus, return hierachical clustering ("clust") or Kmeans ("kmeans") plots with their complete path
}
\description{
Morphometrics classification (groups) between different graphical units (GUs).
 Read JPG files from each different folder. Useful after comparisons (see, m'orph_nds_compar' function)
}
\examples{
morph_nds_group(nodes)

## [1] "* read 'oeil' typo"
## Extracting 10.jpg outlines...
## [ 1 / 10 ]  Ain Ghazal.stat_2.1.jpg
## ...
## [ 10 / 10 ]  Qarassa.figurine__wx.14.jpg

}
