% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_x_nds_morph_classif_resume.R
\name{nds_morph_classif_resume}
\alias{nds_morph_classif_resume}
\title{create a contactsheet (spatialization, kmeans & mean shape, dataframe, ...) for clustered nodes}
\arguments{
\item{nodes}{dataframe of nodes}

\item{gu.types}{classes of nodes that will be clustered, a single value}

\item{mbrshp}{see 'nds_morph_classif()'}

\item{path.coords}{path of the .csv with sites coordinates (for 'spatial' and 'all')}

\item{out.data}{if "all" return a contactsheet with all stats, if not, depend on stats}
}
\description{
create a contactsheet (spatialization, kmeans & mean shape, dataframe, ...) for clustered nodes
}
